+++ autobahn/wamp/component.py
@@ -27,6 +27,7 @@
 
 from __future__ import absolute_import
 
+import itertools
 import six
 import random
 from functools import wraps, partial
@@ -403,13 +404,13 @@ class Component(ObservableMixin):
         """
         self.set_valid_events(
             [
-                'start',        # fired by subclass
+                'start',        # fired by base class
                 'connect',      # fired by ApplicationSession
                 'join',         # fired by ApplicationSession
                 'ready',        # fired by ApplicationSession
                 'leave',        # fired by ApplicationSession
                 'disconnect',   # fired by ApplicationSession
-                'connectfailure',  # fired by subclass
+                'connectfailure',   # fired by base class
             ]
         )
 
@@ -460,11 +461,160 @@ class Component(ObservableMixin):
                 return True
         return False
 
-    def start(self, reactor_or_loop):
-        raise RuntimeError('not implemented')
+    def start(self, loop=None, log_level=None):
+        """
+        This starts the Component, which means it will start connecting
+        (and re-connecting) to its configured transports. A Component
+        runs until it is "done", which means one of:
+
+        - There was a "main" function defined, and it completed successfully;
+        - Something called ``.leave()`` on our session, and we left successfully;
+        - ``.stop()`` was called, and completed successfully;
+        - none of our transports were able to connect successfully (failure);
+
+        :returns: a Future/Deferred which will resolve (to ``None``) when we are
+            "done" or with an error if something went wrong.
+        """
+
+        # Component users self-managing loop & logging need to be able to set
+        # log_level otherwise only info level library log msgs are logged
+        # http://txaio.readthedocs.io/en/latest/programming-guide.html#starting-logging-yourself
+        if log_level is not None:
+            txaio.set_global_log_level(log_level)
+
+        # run() will always pass in a loop, ensure direct callers of this
+        # method do the same
+        if loop is None:
+            self.log.info("No event loop specified")
+            raise RuntimeError("An even loop must be passed to start()")
+
+        # this future will be returned, and thus has the semantics
+        # specified in the docstring.
+        done_f = txaio.create_future()
+
+        # Create a generator of transports that .can_reconnect()
+        transport_gen = itertools.cycle(self._transports)
+
+        # this is a 1-element list so we can set it from closures in
+        # this function
+        transport_candidate = [0]
+
+        def error(fail):
+            self.log.info("Internal error {msg}", msg=txaio.failure_message(fail))
+            self.log.debug("{tb}", tb=txaio.failure_format_traceback(fail))
+            txaio.reject(done_f, fail)
+
+        def attempt_connect(_):
+
+            def handle_connect_error(fail):
+                unrecoverable_error = False
+
+                # FIXME - make txaio friendly
+                # Can connect_f ever be in a cancelled state?
+                # if txaio.using_asyncio and isinstance(fail.value, asyncio.CancelledError):
+                #     unrecoverable_error = True
+
+                self.log.debug(u'component failed: {error}', error=txaio.failure_message(fail))
+                self.log.debug(u'{tb}', tb=txaio.failure_format_traceback(fail))
+                # If this is a "fatal error" that will never work,
+                # we bail out now
+                if isinstance(fail.value, ApplicationError):
+                    if fail.value.error in [u'wamp.error.no_such_realm']:
+                        unrecoverable_error = True
+                        self.log.error(u"Fatal error, not reconnecting")
+
+                    self.log.error(u"{msg}", msg=fail.value.error_message())
+
+                elif isinstance(fail.value, OSError):
+                    # failed to connect entirely, like nobody
+                    # listening etc.
+                    self.log.info(u"Connection failed: {msg}", msg=txaio.failure_message(fail))
+
+                elif self._is_ssl_error(fail.value): #     # Quoting pyOpenSSL docs: "Whenever
+                    # [SSL.Error] is raised directly, it has a
+                    # list of error messages from the OpenSSL
+                    # error queue, where each item is a tuple
+                    # (lib, function, reason). Here lib, function
+                    # and reason are all strings, describing where
+                    # and what the problem is. See err(3) for more
+                    # information."
+                    self.log.error(u"TLS failure: {reason}", reason=fail.value.args[1])
+                    self.log.error(u"Marking this transport as failed")
+                    transport.failed()
+                else:
+                    self.log.error(
+                        u'Connection failed: {error}',
+                        error=txaio.failure_message(fail),
+                    )
+                    # some types of errors should probably have
+                    # stacktraces logged immediately at error
+                    # level, e.g. SyntaxError?
+                    self.log.debug(u'{tb}', tb=txaio.failure_format_traceback(fail))
+
+                if unrecoverable_error:
+                    txaio.reject(done_f, fail)
+                    return
+
+                txaio.call_later(0, transport_check, None)
+                return
+
+            def notify_connect_error(fail):
+                chain_f = txaio.create_future()
+                handler_f = self.fire('connectfailure', self, fail.value)
+                txaio.add_callbacks(
+                    handler_f,
+                    lambda _: txaio.reject(chain_f, fail),
+                    lambda _: txaio.reject(chain_f, fail)
+                )
+                return chain_f
+
+            def connect_error(fail):
+                notify_f = notify_connect_error(fail)
+                txaio.add_callbacks(notify_f, None, handle_connect_error)
+
+            def session_done(x):
+                txaio.resolve(done_f, None)
+
+            connect_f = self._connect_once(loop, transport_candidate[0])
+            txaio.add_callbacks(connect_f, session_done, connect_error)
+
+        def transport_check(_):
+            self.log.debug('Entering re-connect loop')
+
+            if not self._can_reconnect():
+                err_msg = u"Component failed: Exhausted all transport connect attempts"
+                self.log.info(err_msg)
+                try:
+                    raise RuntimeError(err_msg)
+                except RuntimeError as e:
+                    txaio.reject(done_f, e)
+                    return
+
+            while True:
+
+                transport = next(transport_gen)
+
+                if transport.can_reconnect():
+                    transport_candidate[0] = transport
+                    break
+
+            delay = transport.next_delay()
+            self.log.debug(
+                'trying transport {transport_idx} using connect delay {transport_delay}',
+                transport_idx=transport.idx,
+                transport_delay=delay,
+            )
+
+            delay_f = txaio.sleep(delay)
+            txaio.add_callbacks(delay_f, attempt_connect, error)
+
+        # issue our first event, then start the reconnect loop
+        start_f = self.fire('start', loop, self)
+        txaio.add_callbacks(start_f, transport_check, error)
+        return done_f
 
     def stop(self):
-        raise RuntimeError('not implemented')
+        return self._session.leave()
 
     def _connect_once(self, reactor, transport):
 
@@ -569,6 +719,7 @@ class Component(ObservableMixin):
                 return session
 
         transport.connect_attempts += 1
+
         d = self._connect_transport(reactor, transport, create_session)
 
         def on_connect_sucess(proto):
